Fix screens internal structure
