+++ app/src/main/AndroidManifest.xml
@@ -48,12 +48,12 @@
             android:exported="false"
             android:theme="@style/Theme.MusicMap.NoActionBar" />
         <activity
-            android:name=".screens.main.CameraActivity"
+            android:name=".screens.main.musicmemory.create.CameraActivity"
             android:exported="false"
             android:theme="@style/Theme.MusicMap.NoActionBar"
             />
         <activity
-            android:name=".screens.main.MusicMemoryActivity"
+            android:name=".screens.main.musicmemory.MusicMemoryActivity"
             android:exported="false"
             android:theme="@style/Theme.MusicMap.NoActionBar"
             android:configChanges="orientation|screenSize"
