+++ app/src/main/java/com/example/musicmap/models/SongCount.java
@@ -1,6 +1,4 @@
-package com.example.musicmap.screens.artist;
-
-import com.example.musicmap.models.Song;
+package com.example.musicmap.models;
 
 /**
  * This class is used to group song and number of times it was referenced in music memories.
