+++ app/src/main/java/com/example/musicmap/screens/main/artist/ArtistDataMostPlayedSongsFragment.java
@@ -1,4 +1,4 @@
-package com.example.musicmap.screens.main;
+package com.example.musicmap.screens.main.artist;
 
 import android.app.Activity;
 import android.os.Bundle;
@@ -11,8 +11,9 @@ import android.widget.ListView;
 import androidx.annotation.NonNull;
 
 import com.example.musicmap.R;
+import com.example.musicmap.screens.main.MainFragment;
 import com.example.musicmap.util.adapters.PopularSongsAdapter;
-import com.example.musicmap.screens.artist.SongCount;
+import com.example.musicmap.models.SongCount;
 import com.example.musicmap.models.Artist;
 import com.example.musicmap.firebase.Session;
 import com.example.musicmap.models.User;
