+++ app/src/main/res/layout/activity_camera.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/create_music_memory_backgound"
-    tools:context=".screens.main.PostFragment">
+    tools:context=".screens.main.musicmemory.create.PostFragment">
 
     <androidx.camera.view.PreviewView
         android:id="@+id/cameraPreviewView"
