+++ app/src/main/res/layout/fragmen_search.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.main.FeedFragment">
+    tools:context=".screens.main.feed.FeedFragment">
 
     <SearchView
         android:id="@+id/spotify_search_view"
