+++ app/src/main/res/layout/fragment_artist_data.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_marginTop="@dimen/toolbar_margin"
-        tools:context=".screens.artist.ArtistDataFragment">
+        tools:context=".screens.main.artist.ArtistDataFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -34,7 +34,7 @@
 
             <androidx.fragment.app.FragmentContainerView
                 android:id="@+id/fragmentContainerView"
-                android:name="com.example.musicmap.screens.map.ArtistDataMapFragment"
+                android:name="com.example.musicmap.screens.main.artist.ArtistDataMapFragment"
                 android:layout_width="match_parent"
                 android:padding="5dp"
                 android:layout_height="275dp"
@@ -68,7 +68,7 @@
 
             <androidx.fragment.app.FragmentContainerView
                 android:id="@+id/fragment_view_most_played"
-                android:name="com.example.musicmap.screens.main.ArtistDataMostPlayedSongsFragment"
+                android:name="com.example.musicmap.screens.main.artist.ArtistDataMostPlayedSongsFragment"
                 android:layout_width="match_parent"
                 android:layout_height="275dp"
                 android:layout_marginTop="8dp"
