+++ app/src/main/res/layout/fragment_feed.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.main.FeedFragment">
+    tools:context=".screens.main.feed.FeedFragment">
 
     <ListView
         android:id="@+id/feed_list"
