+++ app/src/main/res/layout/fragment_spotify_widget.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.main.SpotifyWidgetFragment">
+    tools:context=".util.ui.SpotifyWidgetFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
